feat(client): support 32-bit item client ids
Item client ids above 65535 (e.g. the koliseuot task tokens 66142/66143/66144)
rendered the wrong sprite -- the id was truncated to 16 bits (66142 & 0xFFFF =
606, which is what drew). The storage was already 32-bit (Item::m_clientId,
ThingType::m_id, and AppearancesLoader sizes the per-category vectors by maxId
in uint32), but the lookup truncated:

- thingtypemanager: getThingType/rawGetThingType/isValidDatId/
  findItemTypeByClientId took `uint16 id` -> uint32. Item::rawGetThingType()
  passes m_clientId; the uint16 param chopped it to 606 before indexing.
- appearancesloader: `type->m_id = static_cast<uint16>(app.id())` -> app.id()
  (m_id is already uint32; the cast was needlessly truncating).

Widening only; every existing caller passes int/uint16. Covers local render by
client id (UIItem:setItemId); receiving such items over the wire is a separate
concern (ProtocolGame item parse).

Target: OTC

+++ src/client/appearancesloader.cpp
@@ -175,7 +175,7 @@ bool AppearancesLoader::buildThingType(const Appearance& app, ThingCategory cate
 {
     auto type = std::make_shared<ThingType>();
     type->m_null = false;
-    type->m_id = static_cast<uint16>(app.id());
+    type->m_id = app.id(); // 32-bit client id; do NOT truncate to uint16 (ids > 65535 exist)
     type->m_category = category;
 
     // Bytes-typed in proto, std::string in C++ — direct assignment is safe.
